Remove ListHead::operator*
It had unintuitive behaviour (Gave refernce to the pointer instead of
the pointee) and was not even used. And there already was implicit
conversion to pointer as well as ToRawPointer().

It was not replaced by the sanely behaving operator * as list should only
be used with bw objects, and those should not be copied around
unnecessarily.