Replace Bullet/Lone Sprite containers with a more specialized solution
Main advantages are that the container is harder to modify incorrectly
and not having to store the container indices in Sprites/Bullets.

Additionally the Bullet state progression was reorganized to be (hopefully)
a bit more cleaner, and saves now have to use key-value mapping for
serializing Bullet/Sprite pointers, as the containers don't provide
random access.