Fix for sprite leaks
Sprite::SingleDelete was supposed to delete the sprite itself, but it
hasn't since Bullet switched to unique_ptrs.
This commit fixes the remaining places where Sprite::SingleDelete was
expected to delete the sprite and removes the useless function.

b05dba047 fixed the larger leak of units leaking all of their sprites.