Fix build with clang 3.8/libstdc++ 5.1
Not sure what is happening with hooks.. Clang doesn't mangle JumpHook with
__fastcall and has collisions..?
Apparently --std=c++14 isn't supposed to provide strcmpi which is needed
in a few places, so switch to --std=gnu++14 for the time being.
Also changes messy Console::AddCommand to be prettier as there were
issues with the std::bind used by it.