Improve pathing/dodging slightly
Some pixel -> subpixel conversions (The ones that aren't done by bw)
now add 128 to the result, to make algorithms behave more symmetrically
with regards to top/left vs bottom/right.

And if a unit needs to dodge something that is directly diagonally from
it, it now picks the shorter path.