Redo object ids/dat accessing
Now, the objects have a separate class for their type (UnitType,
SpriteType, OrderType, ...), which can be used to access their .dat stats
and other functionality that doesn't require an object to exist.
Also the hardcoded ids have been moved to UnitId, SpriteId, OrderId, ...
namespaces to separate them from the objects.

So this required changing a lot of lines... Did some cleaning as well, as
so much had to be changed anyways.

This removes a bunch of version-specific .dat offsets and uses the
DatTable structure instead.